move the contexts directory out the pages directory and change to using environment variables
